+++ packages/core/src/__tests__/data/angular/two-fors.raw.tsx
@@ -1,0 +1,12 @@
+import { For, useState } from '@builder.io/mitosis';
+
+export default function MyComponent() {
+  const [items] = useState([1, 2, 3]);
+
+  return (
+    <div>
+      <For each={items}>{(item) => <div key={item}>{item}</div>}</For>
+      <For each={items}>{(item) => <div key={item}>{item}</div>}</For>
+    </div>
+  );
+}