+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @builder.io/mitosis-cli
 
+## 0.3.17
+
+### Patch Changes
+
+- Updated dependencies [48f5481]
+  - @builder.io/mitosis@0.3.17
+
+## 0.3.16
+
+### Patch Changes
+
+- Updated dependencies [9abf0ac]
+  - @builder.io/mitosis@0.3.16
+
 ## 0.3.15
 
 ### Patch Changes